Update and rename README.md to EJERCICIO_2.md
Have the function `noIterate(strArr)` take the array of strings stored in strArr,
which will contain only two strings, the first parameter being the string N and the second parameter 
being a string K of some characters, and your goal is to determine the smallest substring of N that contains
all the characters in K. For example: if strArr is `["aaabaaddae", "aed"]` then the smallest substring 
of N that contains the characters `a, e, and d` is `"dae"` located at the end of the string.
So for this example your program should return the string `"dae"`.